+++ samples/client/angular/projects/contact/src/app/app.config.ts
@@ -14,7 +14,7 @@
  limitations under the License.
  */
 
-import { MessageProcessor, V0_9_CATALOG, provideA2UI } from '@a2ui/angular';
+import { MessageProcessor, V0_9_CATALOG, provideA2UI, provideMarkdownRenderer } from '@a2ui/angular';
 
 import { IMAGE_CONFIG } from '@angular/common';
 import {
@@ -25,15 +25,17 @@ import {
 import { provideClientHydration, withEventReplay } from '@angular/platform-browser';
 import { theme } from './theme';
 
+import { renderMarkdown } from '@a2ui/markdown-it';
+
 export const appConfig: ApplicationConfig = {
   providers: [
     provideBrowserGlobalErrorListeners(),
     provideZonelessChangeDetection(),
     provideClientHydration(withEventReplay()),
+    provideMarkdownRenderer(renderMarkdown),
     provideA2UI({
       catalog: V0_9_CATALOG,
       theme: theme,
-
       processor: new MessageProcessor(V0_9_CATALOG),
     }),
     {